ai_server/requirements.txt
fastapi
uvicorn
pytesseract
pillow
joblib
imagehash
python-multipart
